codes/code-image-generator/requirements.txt
beautifulsoup4==4.12.2
imageio==2.27.0
imgaug==0.4.0
imutils==0.5.4
keras==2.14.0
opencv_python==4.6.0.66
pillow==10.3.0
requests==2.21.0
scikit-image==0.21.0
scikit-learn==1.4.2
scipy>=1.10.0
seaborn==0.12.2
spacy==3.2.6
tensorflow==2.14.0
validators==0.18.2
matplotlib
html5lib==1.1
joblib==1.3.2
numpy==1.26.4
pandas==2.2.2
wfdb==4.1.2
pyyaml
qrcode==7.4.2
tqdm==4.67.1